+++ R/topography-files.R
@@ -40,6 +40,11 @@ topo_files_generic <- function(pattern, ...) {
 #'
 #' @examples
 #' gebco14_files()
+gebco19_files <- function(all = FALSE, ...) {
+
+  pattern <- if (all) "www.bodc.ac.uk/.*/GEBCO_15SEC/zip" else "www.bodc.ac.uk/.*/GEBCO_15SEC/zip/GEBCO_2019.nc$"
+  topo_files_generic(pattern)
+}
 gebco14_files <- function(all = FALSE, ...) {
   pattern <- if (all) "www.bodc.ac.uk/gebco/" else "www.bodc.ac.uk/.*/GEBCO_2014_2D.nc$"
   topo_files_generic(pattern)